:truck:rename:#45 LNB 폴더구조 변경(components 폴더 안으로 이동)
